refactoring security files
