+++ app.js
@@ -15,7 +15,7 @@ const cookieParser = require('cookie-parser');
 const config = require('config');
 const utils = require(`${__dirname}/app/components/utils`);
 const packageJson = require(`${__dirname}/package`);
-const Security = require(`${__dirname}/app/components/security`);
+const Security = require(`${__dirname}/app/services/security`);
 const helmet = require('helmet');
 const csrf = require('csurf');
 const declaration = require(`${__dirname}/app/declaration`);
